Improve packer build for Vagrant (#177)
* Reformat

Change formatting so taht future changes will yield smaller diffs.

* Consume config from the checked out repository

By using the config file from the local checkout of the git repository,
the packer build condig and the cloud config are in sync when building a
box.

* Ignore Packer and Vagrant artifacts

* Update to release v0.3.0

Make use of the latest iso release.

* Remove k3s runtime

Certificates will no longer be valid when spawning a vagrant box. By
removing them, k3s will recreated a valid set.

* Add Vagrantfile template

The generated Vagrant box has some shortcommings. Attaching a Vagrantfile
to the box prevents users from addinng those configs themself over and
over again.

* Document how to use

Added a README expalining how this is supposed to be used and what the
limitations are. This includes a Vagrantfile that goes with the usage
example.